// src/app/models/weather.model.ts
import { WeatherDescription, WindData } from './forecast.model';

export interface Weather {
  coord: {
    lon: number;
    lat: number;
  };
  weather: WeatherDescription[];
  base: string;
  main: {
    temp: number;
    feels_like: number;
    temp_min: number;
    temp_max: number;
    pressure: number;
    humidity: number;
  };
  visibility: number;
  wind: WindData;
  clouds: {
    all: number;
  };
  dt: number;
  sys: {
    type: number;
    id: number;
    message?: number;
    country: string;
    sunrise: number;
    sunset: number;
  };
  timezone: number;
  id: number;
  name: string;
  cod: number;
}

export const emptyWeather: Weather = {
  coord: {
    lon: 0,
    lat: 0
  },
  weather: [],
  base: '',
  main: {
    temp: 0,
    feels_like: 0,
    temp_min: 0,
    temp_max: 0,
    pressure: 0,
    humidity: 0
  },
  visibility: 0,
  wind: {
    speed: 0,
    deg: 0
  },
  clouds: {
    all: 0
  },
  dt: 0,
  sys: {
    type: 0,
    id: 0,
    message: 0,
    country: '',
    sunrise: 0,
    sunset: 0
  },
  timezone: 0,
  id: 0,
  name: '',
  cod: 0
};